requirements.txt
openai==0.28.1
python-dotenv==1.0.0
matplotlib==3.7.3
anthropic==0.5.0
streamlit==1.28.2